Pong quality of service

+++ software/src/multiplayer/pong-game/Move.h
@@ -23,7 +23,7 @@ public:
     void read(InputMemoryStream &stream) override;
 
 private:
-    int y;
+    int8_t y;
 };
 
 }
